requirements.txt
mysql-connector-python==8.0.23
protobuf==3.15.0
six==1.15.0
